top: clean up simulator
