--- crates/core/src/topology/simulation/simulator.rs
@@ -1,1 +1,0 @@
-
